+++ Ceres/app/src/main/java/com/sysu/ceres/http/ApiService.java
@@ -110,7 +110,7 @@ public interface ApiService {
 
     // 添加调查问卷问题：http://111.230.13.139:8080/ServerAndDB/addQuestion?sid=2&qid=5&qtype=choice&qtitle=where&answer_a=guanghzou&answer_b=shanghai&answer_c=shenzhen&answer_d=others
     @POST("addQuestion")
-    Observable<Status> addQuestion(@Query("sid") long sid, @Query("qid") long qid,
+    Observable<Status> addQuestion(@Query("sid") int sid, @Query("qid") int qid,
                                        @Query("qtype") String qtype, @Query("qtitle") String qtitle,
                                        @Query("answer_a") String answer_a, @Query("answer_b") String answer_b,
                                        @Query("answer_c") String answer_c, @Query("answer_d") String answer_d);
@@ -120,18 +120,18 @@ public interface ApiService {
 
     // 获取指定Task下的所有调查问卷  http://111.230.13.139:8080/ServerAndDB/getSurveyList?tid=3
     @POST("getSurveyList")
-    Observable<SurveyList> getSurveyList(@Query("tid") long tid);
+    Observable<SurveyList> getSurveyList(@Query("tid") int tid);
 
     // 获取指定调查问卷的所有问题 http://111.230.13.139:8080/ServerAndDB/getSurvey?sid=3
     @POST("getSurvey")
-    Observable<SurveyFull> getQuestionList(@Query("sid") long sid);
+    Observable<SurveyFull> getQuestionList(@Query("sid") int sid);
 
     // 更新调查问卷选项信息 http://111.230.13.139:8080/ServerAndDB/updateAnswers?sid=2&qid=4&answer=A
     @POST("updateAnswers")
-    Observable<Status> updateAnswers(@Query("sid") long sid, @Query("qid") long qid, @Query("answer") String answer);
+    Observable<Status> updateAnswers(@Query("sid") int sid, @Query("qid") int qid, @Query("answer") String answer);
 
     // 获取调查问卷统计数据 http://111.230.13.139:8080/ServerAndDB/getAnswers?sid=3
     @POST("getAnswers")
-    Observable<StatisticList> getStatisticsList(@Query("sid") long sid);
+    Observable<StatisticList> getStatisticsList(@Query("sid") int sid);
 
 }
